use double instead of float and fix rds switch

+++ src/rds.h
@@ -11,7 +11,7 @@
 
 #include <stdint.h>
 
-extern void get_rds_samples(float *buffer, int count);
+extern void get_rds_samples(double *buffer, int count);
 extern void set_rds_pi(uint16_t pi_code);
 extern void set_rds_rt(char *rt);
 extern void set_rds_ps(char *ps);
